refactor(layering): fully evacuate scylladb + odbc pools into their connector crates
Follow the clickhouse precedent: move the connection pool + connection impls out of the
shared db_connection_pool crate into the owning connector crate, so the driver deps leave
db_connection_pool's (and transitively runtime's) graph entirely.

- scylladb: db_connection_pool/src/{scylladbpool.rs, dbconnection/scylladbconn.rs}
  -> connector-scylladb/src/scylladb/{pool.rs, conn.rs}.
- odbc: db_connection_pool/src/{odbcpool.rs, dbconnection/odbcconn.rs}
  -> connector-odbc/src/{odbcpool.rs, odbcconn.rs}.

Both connectors now DROP their db_connection_pool dependency (the DbConnectionPool /
DbConnection traits come from the datafusion-table-providers fork, as in connector-clickhouse).
db_connection_pool sheds the scylla, odbc-api, arrow-odbc, dyn-clone, and sha2 deps + the
odbc/scylladb features; async-stream/tokio stay (postgres), secrecy stays (snowflake).
scylla, odbc-api, and arrow-odbc are now restricted_deps (fully connector-confined).

pool/conn are re-exported as 'pub mod' so clippy's avoid-breaking-exported-api treats their
public API as exported (matching their prior db_connection_pool status). Layering guard green
(131 crates). No behavior change.
