+++ crates/data-connectors/connector-scylladb/Cargo.toml
@@ -23,12 +23,13 @@ license-file.workspace = true
 description = "ScyllaDB data connector for Spice.ai runtime"
 
 [dependencies]
+arrow.workspace = true
+async-stream.workspace = true
 async-trait.workspace = true
 data_components = { path = "../../data_components" }
-arrow.workspace = true
 datafusion.workspace = true
 datafusion-table-providers = { workspace = true, features = ["federation"] }
-db_connection_pool = { path = "../../db_connection_pool", features = ["scylladb"] }
+futures.workspace = true
 linkme.workspace = true
 ns_lookup = { path = "../../ns_lookup" }
 paste.workspace = true
