+++ crates/db_connection_pool/src/dbconnection.rs
@@ -14,9 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#[cfg(feature = "odbc")]
-pub mod odbcconn;
-#[cfg(feature = "scylladb")]
-pub mod scylladbconn;
 #[cfg(feature = "snowflake")]
 pub mod snowflakeconn;
